backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.5.0
bcrypt==4.1.2
cryptography==42.0.0
python-dotenv==1.0.0
Flask-SocketIO==5.3.6
python-socketio==5.11.0
python-jose==3.3.0
qrcode==7.4.2
Pillow>=10.2.0
pyzbar==0.1.9
python-multipart==0.0.9
redis==5.0.1
celery==5.3.4
gunicorn==21.2.0
eventlet==0.33.3
Flask-Limiter==3.5.0
bleach==6.1.0
email-validator==2.1.0